requirements.txt
aiohttp==3.8.1
aiohttp-retry==2.4.6
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.1
atpublic==2.3
attrs==21.2.0
backcall==0.2.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cryptography==36.0.0
cycler==0.11.0
decorator==5.1.0
dictdiffer==0.9.0
diskcache==5.2.1
distro==1.6.0
dpath==2.0.5
dulwich==0.20.26
dvc==2.8.3
flatten-dict==0.4.2
flufl.lock==6.0
fonttools==4.28.1
frozenlist==1.2.0
fsspec==2021.11.0
ftfy==6.0.3
funcy==1.16
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
grandalf==0.6
httplib2==0.20.2
idna==3.3
importlib-resources==5.4.0
ipython==7.29.0
jedi==0.18.1
joblib==1.1.0
kiwisolver==1.3.2
mailchecker==4.1.0
matplotlib==3.5.0
matplotlib-inline==0.1.3
multidict==5.2.0
nanotime==0.5.2
networkx==2.6.3
numpy==1.21.4
oauth2client==4.1.3
packaging==21.3
pandas==1.3.4
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
phonenumbers==8.12.37
pickleshare==0.7.5
Pillow==8.4.0
ply==3.11
prompt-toolkit==3.0.22
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
PyDrive2==1.10.0
pygit2==1.7.1
Pygments==2.10.0
pygtrie==2.4.2
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-benedict==0.24.3
python-dateutil==2.8.2
python-fsutil==0.5.0
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rich==10.14.0
rsa==4.7.2
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.1
scipy==1.7.2
seaborn==0.11.2
setuptools-scm==6.3.2
shortuuid==1.0.8
shtab==1.5.0
six==1.16.0
smmap==5.0.0
tabulate==0.8.9
text-unidecode==1.3
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
voluptuous==0.12.2
wcwidth==0.2.5
xgboost==1.5.0
xmltodict==0.12.0
yarl==1.7.2
zc.lockfile==2.0
zipp==3.6.0
